// frontend/src/shared/api/admin/model/documentStatus.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * AI-Consultant Admin API
 * AI-Consultant Admin API
 * OpenAPI spec version: 1.0.0
 */

export type DocumentStatus =
  (typeof DocumentStatus)[keyof typeof DocumentStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DocumentStatus = {
  processing: "processing",
  done: "done",
  failed: "failed",
} as const;