// frontend/src/shared/api/admin/model/documentType.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * AI-Consultant Admin API
 * AI-Consultant Admin API
 * OpenAPI spec version: 1.0.0
 */

export type DocumentType = (typeof DocumentType)[keyof typeof DocumentType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DocumentType = {
  pdf: "pdf",
  markdown: "markdown",
  csv: "csv",
} as const;